+++ lib/periodic.dart
@@ -3,6 +3,7 @@ library periodic;
 export 'src/data/common.dart';
 export 'src/data/config/item_config.dart';
 export 'src/data/config/periodic_config.dart';
+export 'src/data/controller.dart';
 export 'src/data/periodic_data/periodic_data.dart';
 export 'src/periodic_dialog.dart';
 export 'src/periodic_widget.dart';