+++ lib/src/data/periodic_state.dart
@@ -1,4 +1,4 @@
-import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:hooks_riverpod/all.dart';
 import 'common.dart';
 import 'periodic_data/periodic_data.dart';