+++ CHANGELOG.md
@@ -1,3 +1,5 @@
-## [0.0.1] - TODO: Add release date.
+## [1.0.0] - Release.
 
-* TODO: Describe initial release.
+* Creates an interval for periodicity and schedules
+* Periodic widget
+* Periodic dialog